feat: 로그인 상태에 따른 헤더 상태 변화 추가 (#28)
