+++ package.json
@@ -53,3 +53,4 @@
     "prettier": "^3.1.0"
   }
 }
+
