+++ src/redux/modules/authSlice.js
@@ -8,7 +8,7 @@ const initialState = {
   photoURL: localStorage.getItem('photoURL')
 };
 
-const AuthSlice = createSlice({
+const authSlice = createSlice({
   name: 'auth',
   initialState,
   reducers: {
@@ -35,6 +35,6 @@ const AuthSlice = createSlice({
   }
 });
 
-export const { login, logout, updateNickname } = AuthSlice.actions;
-export default AuthSlice.reducer;
+export const { login, logout, updateNickname } = authSlice.actions;
+export default authSlice.reducer;
